third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-grid-expected.txt

This is a testharness.js-based test.
FAIL Grid assert_equals: height expected "90px" but got "180px"
FAIL Inline grid assert_equals: height expected "90px" but got "180px"
Harness: the test ran to completion.
